fix Xsens driver on older kernels
Asguard still runs on a 2.6.23, and the FTDI driver on there seems to not
support the TIOCMGET and TIOCMSET ioctls (the same driver works flawlessly on
newer kernels).

Just disable the usage of those ioctls: the comments suggest it is used by Xsens
Bus Master devices and calibration devices, which we don't need.